+++ src/info/info_cli_parser.cpp
@@ -40,6 +40,7 @@ info_cli_parser_c::init_parser() {
   OPT("t|track-info",    set_track_info,    YT("Show statistics for each track in verbose mode."));
   OPT("x|hexdump",       set_hexdump,       YT("Show the first 16 bytes of each frame as a hex dump."));
   OPT("X|full-hexdump",  set_full_hexdump,  YT("Show all bytes of each frame as a hex dump."));
+  OPT("P|positions",     set_dec_positions, YT("Show the position of each element in decimal."));
   OPT("p|hex-positions", set_hex_positions, YT("Show the position of each element in hexadecimal."));
   OPT("z|size",          set_size,          YT("Show the size of each element including its header."));
 
@@ -97,6 +98,11 @@ info_cli_parser_c::set_file_name() {
     mxerror(Y("Only one source file is allowed.\n"));
 
   m_options.m_file_name = m_current_arg;
+}
+
+void
+info_cli_parser_c::set_dec_positions() {
+  m_options.m_hex_positions = false;
 }
 
 void